Windows tweaks

+++ src/Common/FormatTools.cpp
@@ -16,7 +16,7 @@
 namespace Common {
 
 //--------------------------------------------------------------------------------
-std::string get_mining_speed(uint32_t hr) {
+std::string get_mining_speed(uint64_t hr) {
   if (hr>1e9) return (boost::format("%.2f GH/s") % (hr/1e9)).str();
   if (hr>1e6) return (boost::format("%.2f MH/s") % (hr/1e6)).str();
   if (hr>1e3) return (boost::format("%.2f KH/s") % (hr/1e3)).str();
